fix few ui issues
