+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "super-splat",
-    "version": "0.10.1",
+    "version": "0.10.2",
     "author": "PlayCanvas<support@playcanvas.com>",
     "homepage": "https://playcanvas.com",
     "description": "All the splat things",
