+++ src/spinner.ts
@@ -55,48 +55,6 @@ const css = `
         transform: translateY(-50%) rotate(360deg);
     }
 }
-
-.hand-prompt {
-    display: block;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 90px;
-    height: 90px;
-    pointer-events: none;
-    z-index: 9999999999999;
-
-    transform: translate(-50%, -50%);
-    -webkit-animation: prompt-animate 1s ease-in-out 0s infinite alternate;
-    animation: prompt-animate 1s ease-in-out 0s infinite alternate;
-
-    -webkit-filter: drop-shadow(5px 5px 5px #888);
-    filter: drop-shadow(5px 5px 5px #888);
-}
-
-@-webkit-keyframes prompt-animate {
-    0% {
-        -webkit-transform: translate(-125%, -50%);
-        transform: translate(-125%, -50%);
-    }
-
-    100% {
-        -webkit-transform: translate(25%, -50%);
-        transform: translate(25%, -50%);
-    }
-}
-
-@keyframes prompt-animate {
-    0% {
-        -webkit-transform: translate(-125%, -50%);
-        transform: translate(-125%, -50%);
-    }
-
-    100% {
-        -webkit-transform: translate(25%, -50%);
-        transform: translate(25%, -50%);
-    }
-}
 `;
 
 let container: HTMLElement;
@@ -110,7 +68,7 @@ const startSpinner = () => {
     spinner = document.createElement('div');
     spinner.className = 'static-spinner';
     container.appendChild(spinner);
-    document.getElementById('app-container').appendChild(container);
+    document.getElementById('canvas-container')?.appendChild(container);
 };
 
 const stopSpinner = () => {
